backend/user-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
PyJWT==2.8.0
gunicorn==21.2.0